+++ Assets/SRP/Editor/Serializer/SerializerBase.cs
@@ -6,6 +6,10 @@ using UnityEngine;
 
 namespace LoneTower.SRP {
 	public abstract class SerializerBase {
+		protected Type serializeType;
+		public SerializerBase(Type t) {
+			serializeType = t;
+		}
 		public abstract void Serialize(object o, SerializedProperty prop);
 
 		public abstract void SerializeArray(object[] o, SerializedProperty prop);