+++ Assets/SRP/Runtime/IMonoBehaviourBase.cs
@@ -2,9 +2,7 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace LoneTower.SRP {
-	public interface IMonoBehaviourBase {
-		GameObject gameObject { get; }
-		Transform transform { get; }
-	}
+public interface IMonoBehaviourBase {
+	GameObject gameObject { get; }
+	Transform transform { get; }
 }